Do not exit from a bootcmd shell prematurely
The bootcmd for generating netplan or /e/n/i contains 'exit 0' in the
netplan code-path and on bionic exits the shell used to run bootcmds.
Turns on the bootcmd for writing the proxy settings file to
/etc/apt/apt.conf.d/95-juju-proxy-settings is the last bootcmd in
user-data generated by Juju and is never executed on bionic systems as a
result of 'exit 0' in the bootcmd above it.

95-juju-proxy-settings is later written by proxyupdater so ext4 crtime
timestamp for the file are always greater than cloud-init's bootcmd
timestamps (including the one saying that this module has finished
running). In other words, this file is always created after bootcmd
finishes.

This results in apt updates in modules that follow to fail
(specifically cc_package_update_upgrade_install) if only proxied access
to apt mirrors is available (e.g. no default gateway configured). Which
then results in charm failures as they have no repository index. It
seems that some charms have workarounds for this behavior already.

cloud-init uses subprocess.communicate() when it waits for a
child process (/bin/sh in this case) to finish so all file descriptors
are closed in all child processes that could have been forked/exec-ed
from bootcmds and the processes themselves are gone (unless they
daemonized which is not the case). So the issue is in premature exit,
not buffering.

https://bugs.launchpad.net/juju/+bug/1807615